added resukt wrapper

+++ domains/core/entity/with-state.ts
@@ -1,5 +1,6 @@
 import { Observable } from "rxjs";
 import { map, tap } from "rxjs/operators";
+import {FAILURE_MESSAGE, Result} from "utils/result/dto";
 
 import type { StateRecord } from "domains/core/state/record";
 import { Entity } from "domains/core/entity/index";
@@ -7,28 +8,30 @@ import { Entity } from "domains/core/entity/index";
 import { Domains } from "domains/core/domains/application";
 import {IdentifierI} from "utils/unique-id";
 
+type State = Result<Entity<any>, FAILURE_MESSAGE>;
+
 export interface EntityI<T> {
     id: IdentifierI
     ttl: number
-    read(): Observable<Entity<any>>
-    create?: (value: T) => Observable<Entity<any>>
-    update?: (value: Partial<T>) => Observable<Entity<any>>
-    delete?: (id: number|string) => Observable<boolean>
+    read(): Observable<State>
+    create?: (value: T) => Observable<State>
+    update?: (value: Partial<T>) => Observable<State>
+    delete?: (id: number|string) => Observable<Result<boolean, FAILURE_MESSAGE>>
 }
 
 class EntityWithState<EntityT extends Entity<any>, ValueT> {
     private readonly entity: EntityI<ValueT>;
-    private readonly state: StateRecord<EntityT>
+    private readonly state: StateRecord<Result<EntityT, FAILURE_MESSAGE>>
 
     constructor(entity: EntityI<ValueT>) {
         this.entity = entity;
-        this.state = Domains.shared().state<EntityT>(this.entity.id, this.read);
+        this.state = Domains.shared().state<Result<EntityT, FAILURE_MESSAGE>>(this.entity.id, this.read);
     }
 
     private read = () => {
         return this.entity.read().pipe(
             map(it => ({
-                data: it as EntityT,
+                data: it.isSuccessful ? Result.success(it as unknown as EntityT) : Result.failure(it.error),
                 expiration: this.entity.ttl
             }))
         )
@@ -38,9 +41,9 @@ class EntityWithState<EntityT extends Entity<any>, ValueT> {
         return this.entity.id;
     }
 
-    public data(): Observable<ValueT> {
+    public data(): Observable<Result<ValueT, FAILURE_MESSAGE>> {
         return this.state.data().pipe(
-            map(it => it.get())
+            map(it => it.isSuccessful ? Result.success(it.value.get()) : Result.failure(it.error))
         );
     }
 
@@ -50,10 +53,13 @@ class EntityWithState<EntityT extends Entity<any>, ValueT> {
         }
 
         return this.entity.create(value).pipe(
-            tap((it: EntityT) => {
-                this.state.update(it, this.entity.ttl)
+            tap((it) => {
+                if(!it.isSuccessful) {
+                    return;
+                }
+                this.state.update(Result.success(it.value as unknown as EntityT), this.entity.ttl)
             }),
-            map(() => true)
+            map(it => it.isSuccessful ? Result.success(true) : Result.failure(it.error))
         )
     }
 
@@ -63,10 +69,13 @@ class EntityWithState<EntityT extends Entity<any>, ValueT> {
         }
 
         return this.entity.update(value).pipe(
-            tap((it: EntityT) => {
-                this.state.update(it, this.entity.ttl)
+            tap((it) => {
+                if(!it.isSuccessful) {
+                    return;
+                }
+                this.state.update(Result.success(it.value as unknown as EntityT), this.entity.ttl);
             }),
-            map(() => true)
+            map(it => it.isSuccessful ? Result.success(true) : Result.failure(it.error))
         )
     }
 
@@ -76,10 +85,13 @@ class EntityWithState<EntityT extends Entity<any>, ValueT> {
         }
 
         return this.entity.delete(id).pipe(
-            tap(() => {
-                this.state.delete()
+            tap((it) => {
+                if(!it.isSuccessful) {
+                    return;
+                }
+                this.state.delete();
             }),
-            map(() => true)
+            map(it => it.isSuccessful ? Result.success(true) : Result.failure(it.error))
         )
     }
 }
